adk_lab_01/pyproject.toml
[project]
name = "adk-lab-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.20.0",
    "google-genai>=1.53.0",
]
